fix(elasticsearch): Set vertx dependencies as provided
Closes gravitee-io/issues#2305
